Re-structure and extract SelectionBox component
