+++ src/css/base/_global.scss
@@ -25,8 +25,6 @@ h1 {
   text-shadow: 1px 1px 2px $dark-accent;
   font-size: 2.3em;
   margin: 10px 0;
-  position: absolute;
-  top: 0;
 }
 
 h2 {
